Hard code DEFAULT_LOCALE_RESOURCES in rustc_interface
locale_resources will no longer be used at all soon, so no need for
custom drivers to have the ability to define their own locale_resources.
Also hard code report_ice to not use any locale resources. It only emits
diagnostics that have already been ported to inline fluent messages.
